refactor : 이전 PR 리뷰 반영. 카드 컴포넌트 위치 변경
